Updated user.go.

+++ user.go
@@ -11,21 +11,21 @@ type User struct {
 	CreatedAt         time.Time   `json:"created_at"`
 	Email             string      `json:"email"`
 	EventsURL         string      `json:"events_url"`
-	Followers         float64     `json:"followers"`
+	Followers         int64       `json:"followers"`
 	FollowersURL      string      `json:"followers_url"`
-	Following         float64     `json:"following"`
+	Following         int64       `json:"following"`
 	FollowingURL      string      `json:"following_url"`
 	GistsURL          string      `json:"gists_url"`
 	GravatarID        string      `json:"gravatar_id"`
 	Hireable          bool        `json:"hireable"`
 	HtmlURL           string      `json:"html_url"`
-	ID                float64     `json:"id"`
+	ID                int64       `json:"id"`
 	Location          string      `json:"location"`
 	Login             string      `json:"login"`
 	Name              string      `json:"name"`
 	OrganizationsURL  string      `json:"organizations_url"`
-	PublicGists       float64     `json:"public_gists"`
-	PublicRepos       float64     `json:"public_repos"`
+	PublicGists       int64       `json:"public_gists"`
+	PublicRepos       int64       `json:"public_repos"`
 	ReceivedEventsURL string      `json:"received_events_url"`
 	ReposURL          string      `json:"repos_url"`
 	SiteAdmin         bool        `json:"site_admin"`
